feat: created a ConfigUtils class with static helper functions that will be needed all across serverParser and locationParser finished listenParser()
